Redid the command line to have a 'c'/'compile' command and a 'specgen' command that generates code from a c42 specification (and a help command, also invokable via flag).  The default is to compile if no argument is provided.
I am now generating code that compiles for nim, for example:
`con4m specgen tests/spec/s2-sami.c4m --language-nim`

The above will validate that the spec is well-formed, and then use it
to generate nim code to stdout for loading a config to data
structures, with getters and setters.  I may eventually allow options
for tweaking what does or doesn't get generated, but thats not a
priority at the moment.

--output-file will write the generated code to the specified file
  instead.