When running with unknown-flags-ok, I was previously treating:
-tfoo as -t -t -o -o and not -tfoo=bar (assuming tfoo was an unknown
flag). That was an issue for Chalk re-parsing the docker command line
w/ unknown-flag, as We *did* care about the f.  I decided that, while
this could be configurable, unknown arguments just get stuck in argv
anyway, so with unknown flags on, just eat the whole thing in that
case, which gets us the right docker behavior.  This is what was
*actually* happening when I thought the space in "-f bar" was broken,
because I was getting "f: unknown argument".

Along the way, I made all the fields in FlagSpec and CommandSpec
public, because, for some unfathomable reason, if my chalk code calls
newSpecObj() and passes noSpace = false explicitly (or omits it), the
value is somehow false immediately before the function returns, yet
*true* immediately after it returns.  The fix allows chalk to
instantiate the CommandSpec object itself.  Mind blown on that one.