Static openssl on mac (#105)
Figured out some hackery to circumvent cross-project build issues around static linking of architecture specific, pre-built .a files for the mac. Basically, seemed better to make somewhat slim .a's and directly link w/ lto instead of carrying around dylibs. But I need to stick them into both the con4m exe AND the chalk one, but when chalk loaded the con4m init stuff, its pwd was chalk, not con4m. So I had to do some tom-foolery to figure out where the right con4m libs are, and I had to restructure the repo to make sure the deps stay when we download as a dependency.