WIP: Component System WIP: Component system
Most of the first draft is done. The major exception for the component info is the treecheck code for param blocks.
I also have started a bit of the support work that needs to be done, such as removing static lock checks (the dynamic ones need to be redone still).