Reworked to fill in the gaps
Polished, filling in missing pieces, and getting it to work on a test program.

The stack interface had to be massaged. All our stuff currently passes streams. And it now can take a URL instead, though I think that probably needs a tweak, as I'm not testing it yet. Should work fine in the 'use' statement though.

Also, I reworked some tiny bits in the lexer and parser related to newline handling.
- The interface for statements now has some stuff to handle a graceful statement end.
- Sections and parameters no longer REQUIRE braces.
- The pre-check phase has been combined into the post-check phase.
- Any operator tokens will always cause any following whitespace token to gobble up a newline. Meaning, you should *always* be able to add a single newline after any operator (not literal tokens, etc).