Cache per execution context
I (eventually) realized when `chalk gen` didn't work that there were plenty of cases where we'd be running two execution contexts that might share the same base code, leading to type conflicts across the execution environments. To compensate, I moved the component objects into the execution environment.

This was the quick and simple way.

Ideally, We'd centrally cache the output of parse() but either copy the tree or keep a parallel tree of sorts, to avoid too much duplication.