add prompt attribute to component parameters
this will be required for some of the signing improvements in chalk

the idea is that this will allow to add parameters in components
which only auto-compute values via default callback and will
never prompt the user for the value during chalk load.
This way we can store the computer value in chalkmark for easy
on subsequent executions without any user interaction.