ci: add tests workflow (#120)
For now, the tests fail. At least the `nimble build` step of the new
workflow succeeds in CI - it didn't on Linux until the previous
commit [1].

Add only Linux x86_64 for now. On macOS x86_64, the workflow would
error with:

    clang: error: no such file or directory: '/Users/runner/.local/c0/libs/libffi.a'

and on macOS arm64 it would error with:

    ERROR: unable to find archive matching pattern macosx_arm64.tar.xz

because the Nim nightlies repo doesn't have a binary build for that
platform yet.

[1] 2024-04-08, 5e7a21cad233, "fix: make nimble build succeed on Linux"